requirements.txt
boto3<=1.12.33
botocore<=1.15.33           # via boto3, s3transfer
docutils<=0.15.2            # via botocore
jmespath<=0.9.5           # via boto3, botocore
python-dateutil<=2.8.1
s3transfer<=0.3.3        # via boto3
six<=1.14.0               # via python-dateutil